Почистил. Код отлаженный перенес из песочницы в основной код
