+++ tgbot/keyboards/keyboards.py
@@ -1,5 +1,3 @@
-import logging
-
 from aiogram.types import ReplyKeyboardMarkup, KeyboardButton
 from aiogram.types import InlineKeyboardButton
 
@@ -8,9 +6,8 @@ from aiogram.utils.keyboard import InlineKeyboardBuilder
 from tgbot.database.requests import get_categories, get_priorities, get_urls_by_category, get_urls_by_priority
 
 main = ReplyKeyboardMarkup(resize_keyboard=True, input_field_placeholder="Выберите пункт меню...", keyboard=[
-    [KeyboardButton(text="Категории")],
-    [KeyboardButton(text="Приоритеты")],
-    [KeyboardButton(text="Поиск ссылки по тексту")]
+    [KeyboardButton(text="Категории"), KeyboardButton(text="Приоритеты")],
+    [KeyboardButton(text="Добавление новых данных"), KeyboardButton(text="Поиск данных по тексту")]
 ])
 
 
@@ -27,7 +24,7 @@ async def priorities():
     all_priorities = await get_priorities()
     keyboard = InlineKeyboardBuilder()
     for priority in all_priorities:
-        keyboard.add(InlineKeyboardButton(text=f'{priority.priority}', callback_data=f'category_{priority.id}'))
+        keyboard.add(InlineKeyboardButton(text=f'{priority.priority}', callback_data=f'priority_{priority.id}'))
 
     return keyboard.adjust(1).as_markup()
 
@@ -51,13 +48,8 @@ async def urls_by_priority(priority_id):
 
 
 async def urls_to_save(urls):
-    logging.warning(f'urls_to_save. urls: {urls}')
     keyboard = InlineKeyboardBuilder()
     for idx, url in enumerate(urls):
-        logging.warning(f'urls_to_save. url: {url}')
         keyboard.add(InlineKeyboardButton(text=f"Сохранить: {url}", callback_data=f"save_url_{idx}"))
-
-    # Добавить кнопку для отмены сохранения
-    # keyboard.add(InlineKeyboardButton(text="Отмена", callback_data="cancel_save"))
 
     return keyboard.adjust(1).as_markup()
